apigateway/requirements.txt
# Core web frameworks
fastapi
uvicorn

# HTTP request client
requests

# Data validation
pydantic

# Consul client
python-consul
httpx
python-multipart